feat(sessions): add hours support to --since flag (e.g. 2h, 12h)
Previously --since only accepted days (7d, 30d) or bare integers.
Add support for hours (2h, 12h) by parsing as fractional days.

- _parse_since() now returns float to support fractional days
- store.query() since_days type widened from int to float
- Discovery functions (which work in whole days) get max(1, round())
- Help text updated: "e.g. 7d, 30d, 2h"

Useful for checking very recent sessions without --since 1d including
older sessions from earlier in the day.

+++ packages/gptme-sessions/src/gptme_sessions/store.py
@@ -110,7 +110,7 @@ class SessionStore:
         category: str | None = None,
         harness: str | None = None,
         outcome: str | None = None,
-        since_days: int | None = None,
+        since_days: float | None = None,
     ) -> list[SessionRecord]:
         """Filter session records by criteria."""
         records = self.load_all()
